feat: update DTOs to use Integer for IDs in EscalaRequestDTO and remove unused parameter in AvaliadorResponseDTO

+++ src/main/java/br/gov/agu/nutec/msescala/dto/request/EscalaRequestDTO.java
@@ -12,7 +12,7 @@ public record EscalaRequestDTO(
         LocalDate dataFim,
         Uf uf,
         List<Long> orgaoJulgadorIds,
-        List<Long> avaliadorIds,
-        List<Long> pautistaIds
+        List<Integer> avaliadorIds,
+        List<Integer> pautistaIds
 ) {
 }
